+++ CHANGELOG.md
@@ -9,6 +9,7 @@ We used to have incremental versioning before `0.1.0`.
 ### Misc
 
 - Improves tests: now testing magic numbers in more situations
+- Improves tests: now testing more situations with empty base classes
 
 
 ## 0.2.0 aka Revenge of the Async